refactor(dashboard): address review feedback on dependent-filter fix
- Extract `resolveTransitiveParentIds` and `FilterConfigMap` into a new
  `dashboard/components/nativeFilters/dependencyGraph.ts` module so any
  non-hook caller can reuse the traversal and future graph helpers have
  a clear home alongside `hasCircularDependency`.
- Type the `useSelector` call in `useTransitiveParentIds` with `RootState`.
- Narrow `FilterConfigMap` to an inline `{ cascadeParentIds?: string[] }`
  so dividers and chart-customization entries (which may omit the field)
  are also assignable without using `Pick<Filter, ...>`.
- Drop `any` from `state.test.ts` in favor of `DataMaskStateWithId` and
  `FilterConfigMap`.
- Add a test asserting topological order: the closest ancestor's
  `time_range` override wins over further ancestors, which is the
  behavior `mergeExtraFormData` relies on.
