+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterControls/state.ts
@@ -18,57 +18,13 @@
  */
 import { useMemo } from 'react';
 import { shallowEqual, useSelector } from 'react-redux';
-import {
-  DataMaskStateWithId,
-  ensureIsArray,
-  ExtraFormData,
-  Filter,
-} from '@superset-ui/core';
+import { DataMaskStateWithId, ExtraFormData } from '@superset-ui/core';
+import { RootState } from 'src/dashboard/types';
 import { mergeExtraFormData } from '../../utils';
-
-type FilterConfigMap = Record<string, Pick<Filter, 'cascadeParentIds'>>;
-
-/**
- * Resolve the full set of transitive ancestor filter ids for the given filter,
- * in topological order (furthest ancestor first, closest parent last).
- *
- * Dependent filters only list their *direct* parents in `cascadeParentIds`, but
- * the child filter's options and applied extra-form-data must reflect the full
- * chain of upstream filters. A chain A -> B -> C -> D where each level lists
- * only its direct parent used to compute C's options from B alone and D's
- * options from C alone; this meant the child filter's generated
- * `extra_form_data` did not include A's clause even though A is an effective
- * ancestor. That mismatch caused only a subset of filter values to land on
- * charts until the user pressed Apply a second time.
- *
- * Cycles are silently skipped; the filter-config modal enforces acyclicity at
- * save time, but we still defend against malformed config so that a bad entry
- * can never produce an infinite loop here.
- */
-export function resolveTransitiveParentIds(
-  id: string,
-  filterConfig: FilterConfigMap,
-): string[] {
-  const ordered: string[] = [];
-  const visited = new Set<string>([id]);
-
-  const visit = (currentId: string) => {
-    const parents = ensureIsArray(
-      filterConfig[currentId]?.cascadeParentIds,
-    ) as string[];
-    parents.forEach(parentId => {
-      if (visited.has(parentId)) return;
-      visited.add(parentId);
-      // Depth-first so the furthest ancestors are appended before the
-      // closer parents, giving a stable topological order.
-      visit(parentId);
-      ordered.push(parentId);
-    });
-  };
-
-  visit(id);
-  return ordered;
-}
+import {
+  FilterConfigMap,
+  resolveTransitiveParentIds,
+} from '../../dependencyGraph';
 
 /**
  * Resolve the transitive ancestor ids for a given filter from the live
@@ -77,7 +33,7 @@ export function resolveTransitiveParentIds(
  * always agree on which parents count.
  */
 export function useTransitiveParentIds(id: string): string[] {
-  const filterConfig = useSelector<any, FilterConfigMap | undefined>(
+  const filterConfig = useSelector<RootState, FilterConfigMap | undefined>(
     state => state.nativeFilters?.filters,
     shallowEqual,
   );
